pyproject.toml
[tool.poetry]
name = "fastapi-twitter"
version = "0.1.0"
description = ""
authors = ["andrei-samofalov <andrei.e.samofalov@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.95.1"}
uvicorn = "^0.22.0"
alembic = "^1.10.4"
sqlalchemy = "^2.0.11"
asyncpg = "^0.27.0"
loguru = "^0.7.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
loguru = "^0.7.0"
coverage = "^7.2.4"
pytest-flake8 = "^1.1.1"
pytest-coverage = "^0.0"
pytest-black = "^0.3.12"


[tool.poetry.group.linters.dependencies]
flake8 = "^6.0.0"
flake8-pyproject = "^1.2.3"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
#pythonpath = [
#    '.', 'src'
#]
asyncio_mode = "auto"
addopts = [
    '--ignore-glob=*__init__.py', '--ignore-glob=*pgdata*'
]
markers = []

[tool.isort]
profile = 'black'


[tool.black]
line-length = 79
skip-string-normalization = 'True'
force-exclude = '''.*pgdata.*|.*migrations.*|.*tests.*'''


[tool.flake8]
max-line-length = 79
max-doc-length = 79
max-complexity = 10
exclude = [
    'venv', 'dist', 'conf', 'scripts',
    'pgdata', 'migrations', 'tests'
]


[tool.mypy]
ignore_missing_imports = 'True'